Move fetching user details from `getSession` to layout's `load`
Fetching user details in SvelteKit's `getSession` failed because it uses server-side
`fetch`, which doesn't work w/ relative URL.

A simple fix would be to use a full URL in `getSession`. I don't do this since this would
involve hardcoding the domain (which could be avoided but is still inflexible IMO) and establishing
HTTP session.

Instead, fetching user details is moved to the `load` function of the layout. It can run both on the
server and client and works w/ relative URLs. After fetching user details and extracting a user role,
the role is passed into the layout component as a prop for the layout to set the session. (The `MetadataUploadLink` component
then gets the user role from the session to show the upload link to admins only.)